BufferedIO: Add readBuffered, writeBuffered, and nonblocking variants
These will allow hGetBuf* and hPutBuf* to read and write to/from the device
directly without casting the device to FD.

This code does a lot of "state" updating, e.g.:

    (r, buf'') <- copyToBuffer buf' ptr count
    buf''' <- flushWriteBuffer dev buf''
    return (r, buf''')

This is bug-prone.  It'd be nice if GHC had a flag to warn if you
use x when x' is in scope.