Implement hGetBuf* and hPutBuf* using new BufferedIO methods
Fixes #4144 (see 59d4a66 as well)

This commit slightly changes the behavior of hGetBufSome, hGetBufNonBlocking,
and hPutBufNonBlocking.  They are now less greedy about reading or writing all
requested bytes, stopping when the buffer is exhausted rather than making a
system call to try to squeeze out more bytes.  See #5843.

Also, update doc of hGetBufNonBlocking to say that on Windows, it only
works incorrectly "for files".  This is because the network package
should soon have it's own IODevice for sockets, with a correct
implementation of readNonBlocking.